requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.3

# Database - SQLite (dev) or PostgreSQL (prod)
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
alembic==1.13.2

# Cache (no hiredis - C extension causes issues on Windows)
redis==5.0.7
fakeredis==2.23.2

# ML / AI Detection (no version pin - use latest pre-built wheel)
scikit-learn
numpy

# Plagiarism fingerprinting
datasketch==1.6.5

# Storage (optional - falls back to local filesystem)
boto3==1.34.144

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Document parsing
python-docx==1.1.2
PyPDF2==3.0.1
chardet==5.2.0

# Report generation
reportlab==4.2.2

# Async utilities
aiofiles==23.2.1
httpx==0.27.0

# Observability
structlog==24.2.0

# Language detection
langdetect==1.0.9
